requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dotenv==0.19.2
Werkzeug==2.0.3
WTForms==3.0.1
